More Cleanup and Organization
