+++ React.Front/src/Components/MyVideos.css
@@ -1,0 +1,11 @@
+.VideoPlayer {
+
+    width: 100%;
+    height: 100%;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+}
