+++ React.Front/src/Components/Navigation.css
@@ -1,0 +1,11 @@
+
+
+
+.mainNavbar {
+    flex-direction: column;
+    position: relative;
+    display: flex;
+    flex-wrap: nowrap;
+    align-items: flex-end;
+    justify-content: space-evenly;
+}
