+++ React.Front/src/Components/Navigation.js
@@ -2,7 +2,7 @@ import React, { Component, Profiler, useEffect } from 'react';
 
 import { Navbar, Nav } from 'react-bootstrap';
 import { NavLink, Link, useLocation } from 'react-router-dom'
-
+import './Navigation.css';
 export default function Navigation(props) {
 
 
@@ -27,8 +27,8 @@ export default function Navigation(props) {
                     <> <NavLink className="d-inline p-2 bg-dark text-white" to="/users">
                     Users
                 </NavLink>
-                <NavLink className="d-inline p-2 bg-dark text-white" to="/videos">
-                    videos
+                <NavLink className="d-inline p-2 bg-dark text-white" to="/myvideos">
+                   My Videos
                 </NavLink>
         </>
                 : 
@@ -42,7 +42,7 @@ export default function Navigation(props) {
 
     //fill navbar with links and dynamic content. 
     return (
-        <Navbar bg="dark" expand="lg">
+        <Navbar className="mainNavbar" bg="dark" expand="lg">
             <Navbar.Toggle aria-controls="basic-navbar-nav"></Navbar.Toggle>
             <Navbar.Collapse id="basic-navbar-nav">
                 <Nav className="ml-auto">
