+++ React.Front/src/Components/UploadVideo.css
@@ -1,0 +1,2 @@
+body {
+}
