+++ React.Front/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import App from './Components/App';
 import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from './reportWebVitals';
